fixed datatypes

+++ collections/nemo_nlp/nemo_nlp/data/datasets/token_classification.py
@@ -298,7 +298,7 @@ class BertTokenClassificationDataset(Dataset):
     def __getitem__(self, idx):
         return (np.array(self.all_input_ids[idx]),
                 np.array(self.all_segment_ids[idx]),
-                np.array(self.all_input_mask[idx], dtype=np.float32),
+                np.array(self.all_input_mask[idx], dtype=np.long),
                 np.array(self.all_loss_mask[idx]),
                 np.array(self.all_subtokens_mask[idx]),
                 np.array(self.all_labels[idx]))
@@ -350,6 +350,6 @@ class BertTokenClassificationInferDataset(Dataset):
     def __getitem__(self, idx):
         return (np.array(self.all_input_ids[idx]),
                 np.array(self.all_segment_ids[idx]),
-                np.array(self.all_input_mask[idx], dtype=np.float32),
+                np.array(self.all_input_mask[idx], dtype=np.long),
                 np.array(self.all_loss_mask[idx]),
                 np.array(self.all_subtokens_mask[idx]))
